ALFCOM-2784 - FSR overwriting directory.
git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/alfresco/HEAD@13964 7033f55c-6c29-4ad8-afba-765cb616e738

Former-commit-id: 82d2d12dcf47fcbe32078da38f007116031f90d4

+++ root/projects/deployment/source/java/org/alfresco/deployment/impl/server/DeploymentReceiverServiceImpl.java
@@ -273,13 +273,22 @@ public class DeploymentReceiverServiceImpl implements DeploymentReceiverService,
                         case DIR :
                         {
                             File f = deployment.getFileForPath(path);
-                            if (f.exists())
+                            if(f.exists())
                             {
-                                File dest = new File(f.getAbsolutePath() + ".alf");
-                                f.renameTo(dest);
-                                f = deployment.getFileForPath(path);
+                            	if(f.isFile())
+                            	{
+                            		File dest = new File(f.getAbsolutePath() + ".alf");
+                            		f.renameTo(dest);
+                            		f = deployment.getFileForPath(path);
+                            		f.mkdir();
+                            	}
                             }
-                            f.mkdir();
+                            else 
+                            {
+                            	// create a new dir
+                            	f.mkdir();
+                            }
+                           
                             break;
                         }
                         case FILE :
